Bug Fixes...
Monday (13/06/2016); 12:30

+++ ZSProduct/ZSProduct/Modal/AddProducttoListView.cs
@@ -4,7 +4,7 @@ namespace ZSProduct.Modal
     {
         public int qtd;
 
-        public AddProducttoListView(string description, string productCode, uint store, uint stock, string barCode, double price, double pvp2, string reference, uint supplier, string pcu, int Qtd) : base(description, productCode, store, stock, barCode, price, pvp2, reference, supplier, pcu)
+        public AddProducttoListView(string description, string productCode, uint store, double stock, string barCode, double price, double pvp2, string reference, uint supplier, string pcu, int Qtd) : base(description, productCode, store, stock, barCode, price, pvp2, reference, supplier, pcu)
         {
             qtd = Qtd;
         }
